+++ tests/realtikvtest/addindextest2/global_sort_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package addindextest_test
+package addindextest
 
 import (
 	"context"
