+++ tests/realtikvtest/addindextest3/operator_test.go
