+++ tests/realtikvtest/addindextest4/BUILD.bazel
@@ -1,0 +1,95 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "addindextest_test",
+    timeout = "long",
+    srcs = [
+        "add_index_test.go",
+        "concurrent_ddl_test.go",
+        "failpoints_test.go",
+        "global_sort_test.go",
+        "integration_test.go",
+        "main_test.go",
+        "multi_schema_change_test.go",
+        "operator_test.go",
+        "pitr_test.go",
+    ],
+    embed = [":addindextest"],
+    flaky = True,
+    deps = [
+        "//br/pkg/lightning/backend/external",
+        "//br/pkg/lightning/backend/local",
+        "//br/pkg/storage",
+        "//pkg/config",
+        "//pkg/ddl",
+        "//pkg/ddl/copr",
+        "//pkg/ddl/ingest",
+        "//pkg/ddl/testutil",
+        "//pkg/ddl/util/callback",
+        "//pkg/disttask/framework/dispatcher",
+        "//pkg/disttask/framework/proto",
+        "//pkg/disttask/operator",
+        "//pkg/domain",
+        "//pkg/errno",
+        "//pkg/kv",
+        "//pkg/parser/model",
+        "//pkg/sessionctx",
+        "//pkg/sessionctx/variable",
+        "//pkg/table",
+        "//pkg/table/tables",
+        "//pkg/testkit",
+        "//pkg/util/chunk",
+        "//tests/realtikvtest",
+        "@com_github_fsouza_fake_gcs_server//fakestorage",
+        "@com_github_ngaut_pools//:pools",
+        "@com_github_phayes_freeport//:freeport",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+        "@org_golang_x_sync//errgroup",
+    ],
+)
+
+go_test(
+    name = "addindextest3_test",
+    srcs = ["operator_test.go"],
+    deps = [
+        "//pkg/ddl",
+        "//pkg/ddl/copr",
+        "//pkg/ddl/ingest",
+        "//pkg/disttask/operator",
+        "//pkg/domain",
+        "//pkg/kv",
+        "//pkg/parser/model",
+        "//pkg/sessionctx",
+        "//pkg/table",
+        "//pkg/table/tables",
+        "//pkg/testkit",
+        "//pkg/util/chunk",
+        "//tests/realtikvtest",
+        "@com_github_ngaut_pools//:pools",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@org_golang_x_sync//errgroup",
+    ],
+)
+
+go_test(
+    name = "addindextest4_test",
+    srcs = ["ingest_test.go"],
+    deps = [
+        "//br/pkg/lightning/backend/local",
+        "//pkg/ddl",
+        "//pkg/ddl/ingest",
+        "//pkg/ddl/testutil",
+        "//pkg/ddl/util/callback",
+        "//pkg/errno",
+        "//pkg/parser/model",
+        "//pkg/sessionctx/variable",
+        "//pkg/testkit",
+        "//tests/realtikvtest",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+    ],
+)
